client/test: Move helper test file.
